# Conflicts:
#	src/main/java/com/example/final_project_17team/global/controller/PageController.java

+++ src/main/resources/static/js/myPage.js
@@ -1,6 +1,5 @@
 
 document.addEventListener("DOMContentLoaded", async () => {
-    localStorage.setItem('token', "eyJhbGciOiJIUzM4NCJ9.eyJzdWIiOiJ0ZXN0MTIzMTIzIiwiaWF0IjoxNjk0NTg5NDA2LCJleHAiOjE2OTQ2NzU4MDZ9.7-X2j3_p00R10T1BzEMkYkEaRdlXEobeN-Z0JJKSSl_vlu6okSDARj5RmEhXBSpq");
     const httpModule = await import("./service/HttpHandler.js");
     const userModule = await import("./service/User.js");
     const templateModule = await import("./service/Template.js");
@@ -28,14 +27,21 @@ document.addEventListener("DOMContentLoaded", async () => {
         uploadTemplate(User, paint);
     }
     uploadBtn.addEventListener('click', handleUpload);
+
     const handleLeave = (e) => {
         e.preventDefault();
-        User.deleteUser('/profile');
+        const confirmed = confirm("정말 회원 탈퇴하시겠습니까??");
+        if (confirmed) {
+            User.deleteUser('/profile');
+        } else {
+            console.log("회원 탈퇴가 취소되었습니다.");
+        }
     }
     leaveBtn.addEventListener('click', handleLeave);
+
     const handleGoback = (e) => {
         e.preventDefault();
-        window.location.replace('/main');
+        window.location.replace('/matprint/main');
     }
     goBackBtn.addEventListener('click', handleGoback);
     const showProfile = async () => {
@@ -56,7 +62,7 @@ document.addEventListener("DOMContentLoaded", async () => {
 
     const handleLogout = (e) => {
         e.preventDefault();
-        User.logout('/logout');
+        User.logout('/matprint/login');
     }
     logoutBtn.addEventListener('click', handleLogout);
